Remove prompt_windows and term_windows
