--- rocketpool-cli/cli/prompt/prompt_windows.go
@@ -1,8 +1,0 @@
-//go:build windows
-
-package prompt
-
-// Prompt for password input
-func PromptPassword(initialPrompt string, expectedFormat string, incorrectFormatPrompt string) string {
-	return Prompt(initialPrompt, expectedFormat, incorrectFormatPrompt)
-}
