--- rocketpool-cli/cli/term/term_unix.go
@@ -1,15 +1,0 @@
-//go:build !windows
-
-package term
-
-import (
-	"os"
-	"os/exec"
-)
-
-// Clear terminal output
-func Clear() error {
-	cmd := exec.Command("clear")
-	cmd.Stdout = os.Stdout
-	return cmd.Run()
-}
